requirements.txt
PyYAML
Click
Jinja2
jsonschema